requirements.txt
astroid==1.6.1
attrs==17.4.0
colorama==0.3.9
coverage==4.4.2
Django==2.0.1
djangorestframework==3.7.7
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pluggy==0.6.0
py==1.5.2
pylint==1.8.2
pytest==3.3.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytz==2017.3
six==1.11.0
wrapt==1.10.11
yapf==0.20.1
psycopg2==2.7.3.2
